+++ contracts/reflect/src/testing.rs
@@ -1,3 +1,5 @@
+use std::marker::PhantomData;
+
 use crate::msg::{SpecialQuery, SpecialResponse};
 
 use cosmwasm_std::testing::{MockApi, MockQuerier, MockStorage, MOCK_CONTRACT_ADDR};
@@ -7,7 +9,7 @@ use cosmwasm_std::{to_binary, Binary, Coin, ContractResult, OwnedDeps, SystemRes
 /// this uses our CustomQuerier.
 pub fn mock_dependencies_with_custom_querier(
     contract_balance: &[Coin],
-) -> OwnedDeps<MockStorage, MockApi, MockQuerier<SpecialQuery>> {
+) -> OwnedDeps<MockStorage, MockApi, MockQuerier<SpecialQuery>, SpecialQuery> {
     let custom_querier: MockQuerier<SpecialQuery> =
         MockQuerier::new(&[(MOCK_CONTRACT_ADDR, contract_balance)])
             .with_custom_handler(|query| SystemResult::Ok(custom_query_execute(query)));
@@ -15,6 +17,7 @@ pub fn mock_dependencies_with_custom_querier(
         storage: MockStorage::default(),
         api: MockApi::default(),
         querier: custom_querier,
+        custom_query_type: PhantomData,
     }
 }
 
@@ -56,7 +59,7 @@ mod tests {
         }
         .into();
         let wrapper = QuerierWrapper::new(&deps.querier);
-        let response: SpecialResponse = wrapper.custom_query(&req).unwrap();
+        let response: SpecialResponse = wrapper.query(&req).unwrap();
         assert_eq!(response.msg, "FOOD");
     }
 }