+++ packages/std/src/exports.rs
@@ -8,6 +8,7 @@
 //! the contract-specific function pointer. This is done via the `#[entry_point]`
 //! macro attribute from cosmwasm-derive.
 use std::fmt;
+use std::marker::PhantomData;
 use std::vec::Vec;
 
 use schemars::JsonSchema;
@@ -16,6 +17,7 @@ use serde::{de::DeserializeOwned, Serialize};
 use crate::deps::OwnedDeps;
 use crate::imports::{ExternalApi, ExternalQuerier, ExternalStorage};
 use crate::memory::{alloc, consume_region, release_buffer, Region};
+use crate::query::CustomQuery;
 use crate::results::{ContractResult, QueryResponse, Reply, Response};
 use crate::serde::{from_slice, to_vec};
 use crate::types::Env;
@@ -74,9 +76,10 @@ macro_rules! r#try_into_contract_result {
 ///
 /// - `M`: message type for request
 /// - `C`: custom response message type (see CosmosMsg)
+/// - `D`: custom query type (see QueryRequest)
 /// - `E`: error type for responses
-pub fn do_instantiate<M, C, E>(
-    instantiate_fn: &dyn Fn(DepsMut, Env, MessageInfo, M) -> Result<Response<C>, E>,
+pub fn do_instantiate<M, C, D, E>(
+    instantiate_fn: &dyn Fn(DepsMut<D>, Env, MessageInfo, M) -> Result<Response<C>, E>,
     env_ptr: u32,
     info_ptr: u32,
     msg_ptr: u32,
@@ -84,6 +87,7 @@ pub fn do_instantiate<M, C, E>(
 where
     M: DeserializeOwned + JsonSchema,
     C: Serialize + Clone + fmt::Debug + PartialEq + JsonSchema,
+    D: CustomQuery,
     E: ToString,
 {
     let res = _do_instantiate(
@@ -100,9 +104,10 @@ where
 ///
 /// - `M`: message type for request
 /// - `C`: custom response message type (see CosmosMsg)
+/// - `D`: custom query type (see QueryRequest)
 /// - `E`: error type for responses
-pub fn do_execute<M, C, E>(
-    execute_fn: &dyn Fn(DepsMut, Env, MessageInfo, M) -> Result<Response<C>, E>,
+pub fn do_execute<M, C, D, E>(
+    execute_fn: &dyn Fn(DepsMut<D>, Env, MessageInfo, M) -> Result<Response<C>, E>,
     env_ptr: u32,
     info_ptr: u32,
     msg_ptr: u32,
@@ -110,6 +115,7 @@ pub fn do_execute<M, C, E>(
 where
     M: DeserializeOwned + JsonSchema,
     C: Serialize + Clone + fmt::Debug + PartialEq + JsonSchema,
+    D: CustomQuery,
     E: ToString,
 {
     let res = _do_execute(
@@ -126,15 +132,17 @@ where
 ///
 /// - `M`: message type for request
 /// - `C`: custom response message type (see CosmosMsg)
+/// - `D`: custom query type (see QueryRequest)
 /// - `E`: error type for responses
-pub fn do_migrate<M, C, E>(
-    migrate_fn: &dyn Fn(DepsMut, Env, M) -> Result<Response<C>, E>,
+pub fn do_migrate<M, C, D, E>(
+    migrate_fn: &dyn Fn(DepsMut<D>, Env, M) -> Result<Response<C>, E>,
     env_ptr: u32,
     msg_ptr: u32,
 ) -> u32
 where
     M: DeserializeOwned + JsonSchema,
     C: Serialize + Clone + fmt::Debug + PartialEq + JsonSchema,
+    D: CustomQuery,
     E: ToString,
 {
     let res = _do_migrate(migrate_fn, env_ptr as *mut Region, msg_ptr as *mut Region);
@@ -146,15 +154,17 @@ where
 ///
 /// - `M`: message type for request
 /// - `C`: custom response message type (see CosmosMsg)
+/// - `D`: custom query type (see QueryRequest)
 /// - `E`: error type for responses
-pub fn do_sudo<M, C, E>(
-    sudo_fn: &dyn Fn(DepsMut, Env, M) -> Result<Response<C>, E>,
+pub fn do_sudo<M, C, D, E>(
+    sudo_fn: &dyn Fn(DepsMut<D>, Env, M) -> Result<Response<C>, E>,
     env_ptr: u32,
     msg_ptr: u32,
 ) -> u32
 where
     M: DeserializeOwned + JsonSchema,
     C: Serialize + Clone + fmt::Debug + PartialEq + JsonSchema,
+    D: CustomQuery,
     E: ToString,
 {
     let res = _do_sudo(sudo_fn, env_ptr as *mut Region, msg_ptr as *mut Region);
@@ -165,14 +175,16 @@ where
 /// do_reply should be wrapped in an external "C" export, containing a contract-specific function as arg
 /// message body is always `SubcallResult`
 /// - `C`: custom response message type (see CosmosMsg)
+/// - `D`: custom query type (see QueryRequest)
 /// - `E`: error type for responses
-pub fn do_reply<C, E>(
-    reply_fn: &dyn Fn(DepsMut, Env, Reply) -> Result<Response<C>, E>,
+pub fn do_reply<C, D, E>(
+    reply_fn: &dyn Fn(DepsMut<D>, Env, Reply) -> Result<Response<C>, E>,
     env_ptr: u32,
     msg_ptr: u32,
 ) -> u32
 where
     C: Serialize + Clone + fmt::Debug + PartialEq + JsonSchema,
+    D: CustomQuery,
     E: ToString,
 {
     let res = _do_reply(reply_fn, env_ptr as *mut Region, msg_ptr as *mut Region);
@@ -183,14 +195,16 @@ where
 /// do_query should be wrapped in an external "C" export, containing a contract-specific function as arg
 ///
 /// - `M`: message type for request
+/// - `D`: custom query type (see QueryRequest)
 /// - `E`: error type for responses
-pub fn do_query<M, E>(
-    query_fn: &dyn Fn(Deps, Env, M) -> Result<QueryResponse, E>,
+pub fn do_query<M, D, E>(
+    query_fn: &dyn Fn(Deps<D>, Env, M) -> Result<QueryResponse, E>,
     env_ptr: u32,
     msg_ptr: u32,
 ) -> u32
 where
     M: DeserializeOwned + JsonSchema,
+    D: CustomQuery,
     E: ToString,
 {
     let res = _do_query(query_fn, env_ptr as *mut Region, msg_ptr as *mut Region);
@@ -198,8 +212,8 @@ where
     release_buffer(v) as u32
 }
 
-fn _do_instantiate<M, C, E>(
-    instantiate_fn: &dyn Fn(DepsMut, Env, MessageInfo, M) -> Result<Response<C>, E>,
+fn _do_instantiate<M, C, D, E>(
+    instantiate_fn: &dyn Fn(DepsMut<D>, Env, MessageInfo, M) -> Result<Response<C>, E>,
     env_ptr: *mut Region,
     info_ptr: *mut Region,
     msg_ptr: *mut Region,
@@ -207,6 +221,7 @@ fn _do_instantiate<M, C, E>(
 where
     M: DeserializeOwned + JsonSchema,
     C: Serialize + Clone + fmt::Debug + PartialEq + JsonSchema,
+    D: CustomQuery,
     E: ToString,
 {
     let env: Vec<u8> = unsafe { consume_region(env_ptr) };
@@ -221,8 +236,8 @@ where
     instantiate_fn(deps.as_mut(), env, info, msg).into()
 }
 
-fn _do_execute<M, C, E>(
-    execute_fn: &dyn Fn(DepsMut, Env, MessageInfo, M) -> Result<Response<C>, E>,
+fn _do_execute<M, C, D, E>(
+    execute_fn: &dyn Fn(DepsMut<D>, Env, MessageInfo, M) -> Result<Response<C>, E>,
     env_ptr: *mut Region,
     info_ptr: *mut Region,
     msg_ptr: *mut Region,
@@ -230,6 +245,7 @@ fn _do_execute<M, C, E>(
 where
     M: DeserializeOwned + JsonSchema,
     C: Serialize + Clone + fmt::Debug + PartialEq + JsonSchema,
+    D: CustomQuery,
     E: ToString,
 {
     let env: Vec<u8> = unsafe { consume_region(env_ptr) };
@@ -244,14 +260,15 @@ where
     execute_fn(deps.as_mut(), env, info, msg).into()
 }
 
-fn _do_migrate<M, C, E>(
-    migrate_fn: &dyn Fn(DepsMut, Env, M) -> Result<Response<C>, E>,
+fn _do_migrate<M, C, D, E>(
+    migrate_fn: &dyn Fn(DepsMut<D>, Env, M) -> Result<Response<C>, E>,
     env_ptr: *mut Region,
     msg_ptr: *mut Region,
 ) -> ContractResult<Response<C>>
 where
     M: DeserializeOwned + JsonSchema,
     C: Serialize + Clone + fmt::Debug + PartialEq + JsonSchema,
+    D: CustomQuery,
     E: ToString,
 {
     let env: Vec<u8> = unsafe { consume_region(env_ptr) };
@@ -264,14 +281,15 @@ where
     migrate_fn(deps.as_mut(), env, msg).into()
 }
 
-fn _do_sudo<M, C, E>(
-    sudo_fn: &dyn Fn(DepsMut, Env, M) -> Result<Response<C>, E>,
+fn _do_sudo<M, C, D, E>(
+    sudo_fn: &dyn Fn(DepsMut<D>, Env, M) -> Result<Response<C>, E>,
     env_ptr: *mut Region,
     msg_ptr: *mut Region,
 ) -> ContractResult<Response<C>>
 where
     M: DeserializeOwned + JsonSchema,
     C: Serialize + Clone + fmt::Debug + PartialEq + JsonSchema,
+    D: CustomQuery,
     E: ToString,
 {
     let env: Vec<u8> = unsafe { consume_region(env_ptr) };
@@ -284,13 +302,14 @@ where
     sudo_fn(deps.as_mut(), env, msg).into()
 }
 
-fn _do_reply<C, E>(
-    reply_fn: &dyn Fn(DepsMut, Env, Reply) -> Result<Response<C>, E>,
+fn _do_reply<C, D, E>(
+    reply_fn: &dyn Fn(DepsMut<D>, Env, Reply) -> Result<Response<C>, E>,
     env_ptr: *mut Region,
     msg_ptr: *mut Region,
 ) -> ContractResult<Response<C>>
 where
     C: Serialize + Clone + fmt::Debug + PartialEq + JsonSchema,
+    D: CustomQuery,
     E: ToString,
 {
     let env: Vec<u8> = unsafe { consume_region(env_ptr) };
@@ -303,13 +322,14 @@ where
     reply_fn(deps.as_mut(), env, msg).into()
 }
 
-fn _do_query<M, E>(
-    query_fn: &dyn Fn(Deps, Env, M) -> Result<QueryResponse, E>,
+fn _do_query<M, D, E>(
+    query_fn: &dyn Fn(Deps<D>, Env, M) -> Result<QueryResponse, E>,
     env_ptr: *mut Region,
     msg_ptr: *mut Region,
 ) -> ContractResult<QueryResponse>
 where
     M: DeserializeOwned + JsonSchema,
+    D: CustomQuery,
     E: ToString,
 {
     let env: Vec<u8> = unsafe { consume_region(env_ptr) };
@@ -323,10 +343,14 @@ where
 }
 
 /// Makes all bridges to external dependencies (i.e. Wasm imports) that are injected by the VM
-pub(crate) fn make_dependencies() -> OwnedDeps<ExternalStorage, ExternalApi, ExternalQuerier> {
+pub(crate) fn make_dependencies<D>() -> OwnedDeps<ExternalStorage, ExternalApi, ExternalQuerier, D>
+where
+    D: CustomQuery,
+{
     OwnedDeps {
         storage: ExternalStorage::new(),
         api: ExternalApi::new(),
         querier: ExternalQuerier::new(),
+        custom_query_type: PhantomData,
     }
 }